+++ src/main.jsx
@@ -4,7 +4,7 @@ import { RouterProvider } from "react-router-dom";
 import { Provider } from "react-redux";
 import { store } from "@/redux/configureStore.js";
 import { QueryClientProvider } from "@tanstack/react-query";
-import { queryClient } from "./lib/react-query.js";
+import { queryClient } from "./lib/react-query.jsx";
 import App from "./App.jsx";
 import router from "./routes/router.jsx";
 import "./styles/vendors/normalize.css";
